Implement function type support in the parser and AST
- Introduce FunctionType in the AST to represent function signatures with parameters and return types.
- Update the parser to handle function type parsing, including parameter and return type signatures.
- Refactor integer and float type parsing to utilize a more dynamic approach for bit size determination.
- Enhance tests to validate the new function type parsing and ensure correct handling of function signatures.

+++ compiler/internal/parser/typeParser_test.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 )
 
-func assertIntType(t *testing.T, result ast.DataType, expected types.TYPE_NAME, bitSize int, unsigned bool) {
+func assertIntType(t *testing.T, result ast.DataType, expected types.TYPE_NAME, bitSize uint8, unsigned bool) {
 	t.Helper()
 	intType, ok := result.(*ast.IntType)
 	if !ok {
@@ -25,7 +25,7 @@ func assertIntType(t *testing.T, result ast.DataType, expected types.TYPE_NAME, 
 	}
 }
 
-func assertFloatType(t *testing.T, result ast.DataType, expected types.TYPE_NAME, bitSize int) {
+func assertFloatType(t *testing.T, result ast.DataType, expected types.TYPE_NAME, bitSize uint8) {
 	t.Helper()
 	floatType, ok := result.(*ast.FloatType)
 	if !ok {
@@ -54,7 +54,7 @@ func TestParseIntegerType(t *testing.T) {
 		name     string
 		input    string
 		expected types.TYPE_NAME
-		bitSize  int
+		bitSize  uint8
 		unsigned bool
 	}{
 		{"i8", "i8", types.INT8, 8, true},
@@ -90,7 +90,7 @@ func TestParseFloatType(t *testing.T) {
 		name     string
 		input    string
 		expected types.TYPE_NAME
-		bitSize  int
+		bitSize  uint8
 	}{
 		{"f32", "f32", types.FLOAT32, 32},
 		{"f64", "f64", types.FLOAT64, 64},
